feat: add Dockerfile to run server + web-client
The added Dockerfile makes it easy to run the server together with the
web-client in a single process.

Tags
====

Each new release of either the server or the web-client will result in a
new container image, using the following format:

    blendle/automaat:server-vX.Y.Z_web-client-vX.Y.Z

This is a manual process for now, but it will be automated on the CI in
the future. When automating this, there will also be a "development"
release that represents the latest state of the `master` branch:

    blendle/automaat:master

Configuration
=============

The following environment variables can be set to configure the server:

- DATABASE_URL (default: postgres://postgres@localhost)
- SERVER_BIND (server binding address, default: 0.0.0.0:80)
- SERVER_ROOT (place to serve static files from, default: /public)