feat(web): optimize startup queries
If the search bar does not contain a search query during startup (which
is the most common case), then the "total pipelines count" statistic can
be inferred from the total pipelines returned by the `Pipelines::fetch`
call.

This reduces the number of startup queries from 3 to 2.

If there _is_ a query, we can't know for sure if the returned number of
pipelines is equal to the total number of pipelines, so we still do the
third query to fetch the right pipelines count statistic.