fix(server): remove incorrect Content-Length header
This header is incorrectly set by Actix Web, as it represents the length
of the content _before_ it is compressed.

This results in browsers not caching the file contents, causing in
slower startup times of the web client.

This has been reported upstream, but I haven't found the time to dig
into this. For now, removing the bad header makes the caching work.

See: https://github.com/actix/actix-web/issues/930