feat(server): add variable value selection constraints
If a `Variable` has its `selection_constraint` value set, then any
`VariableValue` that matches the key of that variable will have to have
its value match one of the strings in the constraint set.

For example, if the selection constraint is ["hello", "world"], then a
variable value of "universe" is invalid and will return an error,
whereas a value of either "hello" or "world" is valid.

Clients can use this information to pro-actively restrict the set of
allowed values before submitting them to the server. For example, the
web-client can show a select box, instead of an input field, if a
selection constraint is set for a variable.

See: https://github.com/blendle/automaat/issues/1