fix(processor): separate GraphQL input objects
GraphQL requires input objects to be different objects from regular
objects, even if their values are currently the same.

This is to prevent "laziness" by making fields optional in one context,
to serve the other context without wanting to create a new object. You
are required to separate these two object types from the start, even if
it results in duplicate code, to prevent running into problems down the
line.

It makes sense, and this fixes that issue for the http-request
processor.