feat(web): add support for variable selection constraint
If a variable has its `selection` constraint set, the web client will
limit the accepted variable values based on this list.

If only a single value is accepted, a checked checkbox is show, that
cannot be unchecked (this will change in the future, when optional
values become a thing).

If exactly two values are accepted, two radio buttons are shown, to
toggle between the two values.

If more than two values are accepted, a select box is shown to be able
to select the required value.

Without any selection constraints, the regular text field is shown, for
free value input.

This builds on top of 4eb32b37e5f93dd0a6ada29cbf75d61ce8eb46bc.

Closes https://github.com/blendle/automaat/issues/1.