feat(server): add support for default variable values
Pipeline variables can now have a _default value_.

This is a purely cosmetic addition. Clients can use this value to
pre-fill variable input fields, or select the correct value from a set
of selection constraints as the default value.

This feature works together with the selection constraints feature (see
https://github.com/blendle/automaat/issues/1), in that if a default
value is provided for a variable, and a selection constraint is provided
as well, the default value has to be part of that constraint in order
for the variable to be created successfully.

See: https://github.com/blendle/automaat/issues/3