feat(web): add support for default variable values
If a variable has a default value configured, that value is either shown
as the default value in the text field, the correct radio button is
selected, or the corresponding select box option is activated.

This builds on top of dcde5bf7b0d869b96d44a98abe2ef3c8c3d2388d.

Closes https://github.com/blendle/automaat/issues/3.