feat: compile Automaat using the stable Rust compiler
With the recent web-client refactoring in
2d9bc81bec96ae606c2680090469376f3926b68a, and the switch from the Rocket
framework to Actix for the server in
7a61ed2b93a7486448f6f76a858d3097dd8f0f25, there's nothing blocking us
from moving to the stable release channel of Rust.

From here on out, this project is locked to stable, given that it'll be
used in production. Any changes requiring nightly features of Rust will
have to wait until those features are stabilised.