feat(server): add support for variable example values
An "example value" is purely meant as a visual aid for clients to help
people understand what variable value to provide.

For example, a variable named "First and Last Name" could have an
example value set to "John Doe". The clients must **not** use this as a
default value (for that, see the `defaultValue` attribute of the server
API).

For more details, see: https://github.com/blendle/automaat/issues/7