feat(server): support multiple step output formats
Processors return their output as a string, and this string can
optionally be formatted using markdown.

Before, the server would return the string as-is, and the clients would
have to decide on either showing the raw string, or parsing the Markdown
into some different format (such as HTML).

This required extra logic on the clients, and especially for the
web-client causes a 50% increase in Wasm binary size due to the
requirement of Unicode character tables in the binary.

By moving this logic to the server, the clients can request whichever
format they need, without having to do any extra parsing.

Partially resolves https://github.com/blendle/automaat/issues/29,
although client support is still pending.