feat(web): support deep linking to search queries
Direct linking to a search query using the `?search=<query>` URL query
parameter is now supported.

Furthermore, typing in the search field updates the search query
parameter live, so it can be copied and shared.

This is part of an ongoing effort to support more deep links, as
described in https://github.com/blendle/automaat/issues/14.