feat(web): support deep linking to tasks with values
Direct linking to a task with pre-set variable values is now supported.

For each variable, if the query parameter key matches the variable name,
then the parameter value is used as the default variable value.

Furthermore, changing the values of the variable fields updates the
query parameters live, so it can be copied and shared.

When exiting the task view back to the home page, the query parameters
are removed from the URL, to prevent interference between different
tasks using the same variable names.

This is part of an ongoing effort to support more deep links, as
described in #14.