feat(server): encrypt all variable values
Before, when a job was created from a task, the variable values would
get injected into the processor configuration before the configuration
was saved to the database in a JSON blob.

With this change, variable values of a job are now stored in a separate
table, and the values are encrypted using Postgres' `pgcrypto` module.

The variable values are decrypted at runtime, right before they are
injected into the processor configuration, which from that point on only
lives in memory.

WARNING: The server uses a default encryption secret "default secret" to
make it easier to run things locally. This will change before a 1.0
release, and you should always set the secret using the `SERVER_SECRET`
environment variable when running Automaat in production.

This closes https://github.com/blendle/automaat/issues/30.

This also has the added benefit that variable substitution is moved to a
single place in the code base, making it easier to implement
https://github.com/blendle/automaat/issues/23 in the future.

Furthermore, the implemented encryption strategy will make it easier to
implement https://github.com/blendle/automaat/issues/24.