feat(server): add support for (encrypted) global variables
It is now possible to create variables separately from tasks. These
variables are scoped globally, and are always available in task
templates.

Similar to job variables, global variables are encrypted at rest (see
015c41d6d6bb2157a80a4acc3c6c9d2b9a1a43c5).

The GraphQL API has support for creating variables, and optionally
updating existing variables. It does not support fetching the values of
existing global variables, partially because there is no practical need
for this, and partially to improve security.

Closes https://github.com/blendle/automaat/issues/24.