feat(server): support for variable advertisements
This is an initial implementation of the concept of "variable
advertisements".

Check out https://github.com/blendle/automaat/issues/20 for a full
description on what this enables, but the TL;DR is this:

If you have a task `A` that requires a variable `X`, and a task `B`
which has a step defined that returns variable `X`, you can now
"advertise" that task `B` can provide a value for any other task that
requires variable `X` as its input.

Once implemented, this allows the client to show a message when someone
wants to run task `A`, but they don't know the value of `X`, telling
them that they might be able to get that value by running task `B`.

A more advanced client-side implementation could automate this process,
by allowing someone to run task `B`, take the output of `X`
programatically, and use that value as the input of `X` for task `A`.