feat(web): cross-task variable advertisement links
When a task advertises its capability to provide a value for a variable,
any other task requiring that variable value as its input will add a
link to the advertising task.

If more than one task advertises this capability for the same variable
name, a generated dropdown shows each task's title, and their
description, allowing the person that needs the value for a variable to
make an educated choice which task to choose for their value.

As an example, if a task requires the UUID of a customer, one task might
be able to provide this UUID by searching for an email address, and
another might give the UUID based on a first and last name. The person
needing the customer UUID can them choose between those two advertisers,
based on the customer data they have at hand.

This is a first step towards solving
https://github.com/blendle/automaat/issues/20.

For this first implementation, there is no other automation, other than
the automated links added to the relevant advertising tasks.