feat(web): support multiple active tasks
It is now possible to move from one task to another, and have the "back"
button move you back to the previous task, instead of the home page.

Internally, instead of having a single "active task", we now keep track
of a list of active tasks instead, every time a new task is opened, it
is pushed onto the stack of active tasks.

When the back button is clicked, the latest active task is popped off
the stack.

When returning to the home screen, the list of active tasks is cleared.

This builds on top of the feature introduced in
7a92f9ff47ef695a2b277a108dbcf06dd460c53f.

There is still some work to be done to make this work nicely with tasks
that have active jobs.

Solving that is is part of
https://github.com/blendle/automaat/issues/34.