fix(web): hide link to last non-finished result
When re-entering a task you've viewed before, there will be a link to
show the results of the last run.

This link was shown even if the last registered job was still running,
and clicking it didn't show any results, given that there weren't any
yet for that job.

This is now fixed by only looking at finished jobs when deciding if the
link to the last job results should be shown.