fix(web): correctly render job output as HTML
This swaps out the old hack introduced in
2d9bc81bec96ae606c2680090469376f3926b68a to correctly render job output
as HTML.

The underlying issue of not being able to manipulate the raw HTML in the
virtual DOM still exists, but this is a more elegant work-around for
that issue.

Specifically, the old implementation had some corner-cases when the raw
escaped job output wasn't re-rendered as proper HTML.

With this change, those edge-cases are gone, and there is much less
runtime overhead.