feat(server): make ENCRYPTION_SECRET variable required
The `SERVER_SECRET` environment variable has been renamed to
`ENCRYPTION_SECRET` to better reflect its purpose, and the fact that it
is also required for the worker to function.

Its default value of "default secret" has also been removed. Both the
server and the worker will now refuse to start unless this variable is
set.

The variable is still configured as a global. We might still want to
change this at some point, but this is fine for the time being.