feat(server): add session authentication
With this change, the GraphQL endpoint is put behind an authentication
middleware. You are required to supply the `authorization` header with
the correct authentication token.

Tokens are stored as plain-text UUIDv4 in the database (for now), as
this authentication mechanism is not meant to secure the server from the
outside world, but rather to limit internal company-wide access.

In the future, the new `sessions` table will also be used to store
session-related configuration properties.

To facilitate this change, a new `RequestState` object is introduced
that is created for each incoming request. It contains a database
connection handle to perform the necessary queries, and the `Session`
object that will contain the preferences of the current active session
in the future.

There is also a new `createSession` endpoint, that will generate a new
session key that can be distributed within your organisation. Again, in
the future this will be extended to update session preferences.

There is no default session, so you'll have to create your first session
manually using a query such as `INSERT INTO sessions DEFAULT VALUES;`
(the table is configured to auto-generate UUIDv4 values for tokens).

Client-side support has yet to be implemented.

This is part of an ongoing effort described in
https://github.com/blendle/automaat/issues/19.