feat(web): add support for server authentication
With this change, the web client now supports authentication as
implemented by the server in 411e8e21037fd1bb9ea1ee94de7920b25ae8c4f0.

The GraphqlService now redirects the client to the newly added `Login`
page if the server indicates authentication is required. On this page,
a session token can be provided. Once provided and verified, the page
redirects to the home page of Automaat.

The session token is stored in a secure cookie. Visiting the login page
removes the cookie, acting as both a login and logout page.

It's still a bit rough around the edges, and could use another refactor
or two, but the concept works fine as is.

This is another step towards solving
https://github.com/blendle/automaat/issues/19.

Now that there is a way to fetch and store session data, the path is
clear to introduce persistent state to be able to store favorite tasks,
and implement other functionality that requires state to be stored.