feat(server): add privileges-based layered API access
This commit builds on top of 411e8e21037fd1bb9ea1ee94de7920b25ae8c4f0 to
provide a more layered authentication/authorization model, based on the
thoughts described in https://github.com/blendle/automaat/issues/39.

The goal of this commit is to transform the "all-or-nothing"
authentication model into one that allows for fine-grained access
control _and_ allow unauthenticated access to the system.

This creates a more friendly environment for people within an
organisation using Automaat for the first time. Instead of shoving them
a login dialog in their face on their first visit, the server allows
access to all GraphQL queries, to support a client that can show the
full capabilities of API.

Instead, a new privileges-based system is implemented that restricts
GraphQL mutation access based on a set of matching rules where each
mutation requires a specific privilege to be granted to a session.

For tasks, each task can define one or more labels, for which at least
one must match a session's privileges in order to run that task.

If a task defines no labels, the task can be run by unauthenticated
sessions. This allows for patterns where simple side-effect-free tasks
can be used by anyone in the organisation to retrieve vital information
via Automaat, without having to authenticate.