feat(server): allow creating variables without constraints
Requiring to always supply a constraint object, even if no constraints
are needed for the sake of "future compatibility" is silly.

This makes actual usage in practice a bit better, as you won't get
yelled at for not adding an object that has no reason for being there in
the first place.