feat(web): allow closing tasks with running jobs
With recent improvements in the application, such as the introduction of
multiple workers (6de4c96eaf8c2617a801cf3b54526bad3dcd9867) and fixing
the HTML rendering of job results
(20c20f42a13aa3f14d190337c99c8cffa54ab724), it is now possible to close
a task that has an actively running job.

This change makes it so that the "back" button is no longer disabled
when a job is running, and it keeps the "Run" button disabled for as
long as that job is running, even if you closed and re-opened the task
in-between.

It is still not possible (at least not in the UI, the API does support
this) to run more than one job of the same task simultaneously.

This closes https://github.com/blendle/automaat/issues/34.