fix(web): make long task output readable
Before, long task output would flow off the screen without a way to see
the content.

With this change, the inner task details view becomes scrollable when
more output is shown than what fits on the screen.

The "variable advertisers" popup had to be tweaked a bit to make it
not trigger a scrollbar but hover on top of the task details window.