feat(processor): add support for SQL statement parameters
Before this commit, you had to inject variables in statements using
templating, such as:

    "SELECT * FROM table WHERE name = '{{ var['Name'] }}'"

This was prone to SQL injection attacks, and even though Automaat is
supposed to run in your internal network, this is still not great to
expose people to.

With this commit, you can now define statement parameters to be injected
in a safe manner:

    "SELECT * FROM table WHERE name = $1"

You can define the "SQL Query" processor like so:

    SqlQuery {
        statement: "SELECT * FROM table WHERE name = $1 OR id = $2".to_owned(),
        parameters: vec![Type::text("{{ var['Name'] }}"), Type::int(2)],
        url: "postgres://postgres@127.0.0.1".to_owned(),
    };

This commit also updates the version of the `juniper` crate, to support
the needed Rust newtype pattern in the GraphQL schema.