feat(server): provide access to output of all steps
Before this commit, you could access the output of the last step that
ran using the `sys['previous step output']` variable. This was both
restrictive (you couldn't store that output for use in a step further
along the task), and error prone (re-ordering steps would change the
semantics of what that variable would contain).

With this commit, these two problems are resolved with the following
changes:

* You can access the output of all completed steps using the
  `output['<step name>']` syntax.

* You can access task context-specific data using the `context` map
  (which right now exposes the `workspace path` key).

* The top-level `sys` map is removed (it used to contain two values, the
  `previous step output`, which is replaced by the new `output` map, and
  `workspace path`, which is moved to the `context` map).

If two or more steps share the same name, the output key of that name
will contain the output of the last completed step.

Closes https://github.com/blendle/automaat/issues/22.