fix(server): allow tasks without variables
The documentation already stated the original intent of allowing tasks
without variables to be valid. The code did not reflect that intent.

This corrects that mistake.

In reality, there are few cases that don't require a variable, but they
exist, and there is no reason to make it more cumbersome to write them
by having to provide an empty array as the variables value.