refactor(web): extract job result component
The part that shows the job result is getting more complex, and will get
more features added in the future (such as a button to copy the results
to your clipboard).

This refactor extracts the "job result" logic into its own component,
making it easier to reason about in preparation for those upcoming
improvements to this part of the web client.