feat(server): support updating existing tasks
This change introduces the possibility to update tasks.

The implementation is similar to global variables introduced in
3932d5595a5705e3117d9f3535660cd912c066bc:

A new `onConflict` directive can be passed along with the task details.
It can be either `UPDATE` or `ABORT`, it defaults to `ABORT` if unset.

If set to `ABORT`, the behavior will be similar to the existing
behavior, and an error is returned if a similarly named task already
exists.

If set to `UPDATE`, the following semantics apply:

* If a task with a similar name **does not** yet exist, a new task is
  created (this also means that task names currently can't be renamed
  via the API, but they can be modified in the database directly).

* If a similarly named task name **does exist**:
    * its properties (description, labels) are updated,
    * its steps are matched using their names (you can update step
      names, but it will remove the old step, and create a new step,
      instead of updating the existing one),
    * step properties (description, processor, position,
      advertisedVariableKey) are updated,
    * missing steps are removed,
    * new steps are added,
    * the same rules apply to task variables, you can update their
      description, selectionConstraint, defaultValue and exampleValue.

There is no support in the client(s) yet for creating or updating tasks,
but using the exposed GraphQL API playground, you can achieve the same
results.